changes made

+++ Site.Server/Repositories/CartRepository.cs
@@ -8,14 +8,16 @@ namespace Site.Server.Repositories
 {
     public class CartRepository : ICartRepository
     {
-        private readonly CartDatabaseContext _context = null;
-        private readonly IImageRepository _imageRepository = null;
-        private Dictionary<string, ImageModel[]> allImages = null;
+        private CartDatabaseContext? _context = null;
+        private readonly IImageRepository? _imageRepository = null;
+        private Dictionary<string, ImageModel[]>? allImages = null;
         public CartRepository(CartDatabaseContext context, IImageRepository imageRepository)
         {
             _context = context;
             _imageRepository = imageRepository;
         }
+
+
 
         private async Task InitializeImageDictonary()
         {
@@ -23,23 +25,29 @@ namespace Site.Server.Repositories
             string[] names = { "pants", "shirt", "shoe", "shoes" };
 
 
-            foreach (string name in names)
+            if (_imageRepository is not null)
             {
-                ImageModel[] images = await _imageRepository.GetImages(name);
-                allImages.Add(name, images);
+                foreach (string name in names)
+                {
+                    ImageModel[] images = await _imageRepository.GetImages(name);
+                    allImages.Add(name, images);
+                }
             }
         }
         // FIXME: STILL TESTING
         public async Task RemoveProductFromCart(ProductModel cartItem)
         {
-            string title = cartItem.Title.Length >= 6 ? cartItem.Title.Substring(0, 6) : cartItem.Title;
-            var product = _context.Carts.FirstOrDefault(x => x.Username.Equals(cartItem.Username) &&
-                                                             x.Title.Contains(title));
-
-            if (product != null)
+            if (_context is not null)
             {
-                _context.Carts.Remove(product);
-                await _context.SaveChangesAsync();
+                string title = cartItem.Title.Length >= 6 ? cartItem.Title.Substring(0, 6) : cartItem.Title;
+                var product = _context.Carts.FirstOrDefault(x => x.Username.Equals(cartItem.Username) &&
+                                                                 x.Title.Contains(title));
+
+                if (product != null)
+                {
+                    _context.Carts.Remove(product);
+                    await _context.SaveChangesAsync();
+                }
             }
         }
 
@@ -51,71 +59,80 @@ namespace Site.Server.Repositories
 
         public async Task UpdateCart(ProductModel cartItem)
         {
-            var product = _context.Carts.FirstOrDefault(x => x.Username.Equals(cartItem.Username) &&
-                                                                      x.Title.Equals(cartItem.Title));
-            
-            if (product != null)
+            if (_context is not null)
             {
-                product.Quantity = cartItem.Quantity;
-                product.Title = cartItem.Title;
-                product.Username = cartItem.Username;
-                product.Price = cartItem.Price;
-                product.Link = cartItem.Link;
+                var product = _context.Carts.FirstOrDefault(x => x.Username.Equals(cartItem.Username) &&
+                                                                          x.Title.Equals(cartItem.Title));
 
-                _context.Carts.Attach(product);
-                _context.Entry(product).State = EntityState.Modified;
+                if (product != null)
+                {
+                    product.Quantity = cartItem.Quantity;
+                    product.Title = cartItem.Title;
+                    product.Username = cartItem.Username;
+                    product.Price = cartItem.Price;
+                    product.Link = cartItem.Link;
+
+                    _context.Carts.Attach(product);
+                    _context.Entry(product).State = EntityState.Modified;
 
 
-                await _context.SaveChangesAsync();
+                    await _context.SaveChangesAsync();
+                }
             }
-
         }
 
         public async Task FilterCarts()
         {
-            var filterQuery = @"
-                with cte as (
-	                select id, username, title, price, quantity, link, row_number() over 
-	                (partition by username, title order by id) rn
-	                from carts
-                )
-                delete from cte where rn > 1;
-            ";
+            if (_context is not null)
+            {
+                var filterQuery = @"
+                    with cte as (
+	                    select id, username, title, price, quantity, link, row_number() over 
+	                    (partition by username, title order by id) rn
+	                    from carts
+                    )
+                    delete from cte where rn > 1;
+                ";
 
-            await _context.Database.ExecuteSqlRawAsync(filterQuery);
-            await _context.SaveChangesAsync();
+                await _context.Database.ExecuteSqlRawAsync(filterQuery);
+                await _context.SaveChangesAsync();
+            }
         }
 
 
         public async Task<List<ProductModel>> GetAllProducts()
         {
-            var productsFromDb = await _context.Carts.ToListAsync();
             List<ProductModel> models = new List<ProductModel>();
-
-            foreach (Products product in productsFromDb)
+            if (_context is not null)
             {
-                var existingCartItems = models.Where(x => x.Username.Equals(product.Username) &&
-                                                          x.Title.Equals(product.Title)).ToList();
+                var productsFromDb = await _context.Carts.ToListAsync();
+                models = new List<ProductModel>();
 
-                if (!existingCartItems.Any())
+                foreach (Products product in productsFromDb)
                 {
-                    models.Add(
-                    new ProductModel
+                    var existingCartItems = models.Where(x => x.Username.Equals(product.Username) &&
+                                                              x.Title.Equals(product.Title)).ToList();
+
+                    if (!existingCartItems.Any())
                     {
-                        Title = product.Title,
-                        Link = product.Link,
-                        Price = product.Price,
-                        Username = product.Username,
-                        Quantity = product.Quantity
-                    });
+                        models.Add(
+                        new ProductModel
+                        {
+                            Title = product.Title,
+                            Link = product.Link,
+                            Price = product.Price,
+                            Username = product.Username,
+                            Quantity = product.Quantity
+                        });
+                    }
+                    else
+                    {
+                        // Find the item and increment the quantity
+                        ProductModel existingItem = existingCartItems.First();
+                        existingItem.Quantity += 1;
+                    }
+
                 }
-                else
-                {
-                    // Find the item and increment the quantity
-                    ProductModel existingItem = existingCartItems.First();
-                    existingItem.Quantity += 1;
-                }
-                
             }
             return models;
         }
@@ -123,25 +140,28 @@ namespace Site.Server.Repositories
         public async Task AddToCart(CartItemModel cartItem)
         {
             await InitializeImageDictonary();
-            var productsFromDb = await _context.Carts.ToListAsync();
-
-            if (allImages != null)
+            if (_context is not null)
             {
-                foreach (var entry in allImages)
+                var productsFromDb = await _context.Carts.ToListAsync();
+
+                if (allImages != null)
                 {
-                    ImageModel? match = entry.Value.Where(x => x.Title.Contains(cartItem.Title.Substring(0, 5))).ToList().FirstOrDefault();
-                    if (match != null)
+                    foreach (var entry in allImages)
                     {
-                        Products products = new Products()
+                        ImageModel? match = entry.Value.Where(x => x.Title.Contains(cartItem.Title.Substring(0, 5))).ToList().FirstOrDefault();
+                        if (match != null)
                         {
-                            Username = cartItem.Username,
-                            Title = match.Title,
-                            Link = match.Link,
-                            Price = 10.00
-                        };
-                        await _context.Carts.AddAsync(products);
-                        await _context.SaveChangesAsync();
-                        break;
+                            Products products = new Products()
+                            {
+                                Username = cartItem.Username,
+                                Title = match.Title,
+                                Link = match.Link,
+                                Price = 10.00
+                            };
+                            await _context.Carts.AddAsync(products);
+                            await _context.SaveChangesAsync();
+                            break;
+                        }
                     }
                 }
             }
